+++ instrumentation/opentelemetry-instrumentation-django/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Django instrumentation is now enabled by default but can be disabled by setting `OTEL_PYTHON_DJANGO_INSTRUMENT` to `False` ([#1239](https://github.com/open-telemetry/opentelemetry-python/pull/1239))
+- Bugfix use request.path replace request.get_full_path(). It will get correct span name ([#1309](https://github.com/open-telemetry/opentelemetry-python/pull/1309#))
 
 ## Version 0.14b0